Clean AtGRAWUnpacker
Will only search for number of events if it is asked (instead of in
Init)
Break repeated code out into functions (should also break the
processing of a BasicFrame out but it is late and I am tired)
Verified it now works with GADGET test macros still.
Fix AtGRAWUnpacker so FPN channels are not hard-codded. It uses AtMap
instead (possibly relevant for SpecMAT)