Can now change branches through GUI
This also introduces the data handling/update system for tabs. At the
moment it is fairly simple, just a straight broadcast. If more
granularity is needed in the future additional data communication
classes can be added as well as additional AtTabInfo classes that are
more discerning in what updates they accept.