Reduce AtTabBase and refactor AtTabTask
The user never needs to know about AtTabTask and them having access
could cause problems (ie for tabs to update properly if some task is
running then it always need to be the last task added to the run). Now
instead of creating tabs and adding them to a task that is added to
the EventManager they are added directly to the event manager.