Add skeleton framework for sidebar.
Added the run info and event number control to the framework. Didn't
have time to add the branch control, but this skeleton doesn't mess
with the data transfer studd at all. That level of abstraction will
have to be added in next. That is added to the sidebar using the old
method AddFrame().

My thought with this is the sidebar class will create the parts that
are always there (run info, event control, and branch control). Users
can then add additional frames through
`AtEventManagerNew::Instance()::GetSidebar()::AddFrame(...)` which
takes either a enum corresponding to a specific AtSidebarFrame or just
any TGFrame if they don't want to mess with implementing a class for
their additions to the sidebar.